LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/reload-rejection-order-pagehide-unserializablestate-expected.txt



FAIL reload() with an unserializable state inside onpagehide throws "DataCloneError", not "InvalidStateError" assert_equals: expected 1 but got 0
